+++ src/Sifo/Cookie.php
@@ -20,8 +20,8 @@
 
 namespace Sifo;
 
-use Sifo\Filter\FilterCookie;
-use Sifo\Filter\FilterCookieRuntime;
+use Sifo\Filter\Cookie as FilterCookie;
+use Sifo\Filter\CookieRuntime;
 
 class Cookie
 {
@@ -58,7 +58,7 @@ class Cookie
 		}
 
 		// Filter runtime update:
-		FilterCookieRuntime::setCookie( $name, $value );
+		CookieRuntime::setCookie( $name, $value );
 
 		return true;
 	}
@@ -74,7 +74,7 @@ class Cookie
 		}
 
 		// Filter runtime update:
-		FilterCookieRuntime::deleteCookie( $name );
+		CookieRuntime::deleteCookie( $name );
 
 		return true;
 	}
@@ -88,7 +88,7 @@ class Cookie
 	 */
 	static public function get( $cookies )
 	{
-		trigger_error( "'Cookie::get' is deprecated, please use 'FilterCookie'. Ex: FilterCookie::getInstance()->getString( 'cookie_key' );" );
+		trigger_error( "'Cookie::get' is deprecated, please use 'Sifo\\Filter\\Cookie'. Ex: FilterCookie::getInstance()->getString( 'cookie_key' );" );
 
 		if ( is_array( $cookies ) )
 		{
