+++ src/Sifo/Filter/Custom.php
@@ -20,7 +20,9 @@
 
 namespace Sifo\Filter;
 
-class FilterCookie extends Filter
+use Sifo\Exception\FilterException;
+
+class Custom extends Filter
 {
     /**
      * Singleton object.
@@ -30,13 +32,34 @@ class FilterCookie extends Filter
     protected static $instance;
 
     /**
-     * Filters variables passed inside Cookies.
+     * Allow creation of different objects, the FilterCustom is not based on
+     * global values like $_GET or $_POST and might be used for different purposes
+     * in the same execution thread.
+     *
+     * @param array $request
+     *
+     * @return Custom
+     */
+    public function __construct($request)
+    {
+        return parent::__construct($request);
+    }
+
+    /**
+     * Filters variables passed in the array and empties original input.
+     *
+     * @throws FilterException
      * @return Filter
      */
     public static function getInstance()
     {
+        $params = func_get_args();
+        if ((!isset($params[0])) || (!is_array($params[0]))) {
+            throw new FilterException('The variable passed inside the getInstance( $array ) method is not an array.');
+        }
+        $array = $params[0];
         if (!self::$instance) {
-            self::$instance = new self($_COOKIE);
+            self::$instance = new self($array);
         }
 
         return self::$instance;
