+++ src/Sifo/Filter/Request.php
@@ -20,7 +20,7 @@
 
 namespace Sifo\Filter;
 
-class FilterServer extends Filter
+class Request extends Filter
 {
     /**
      * Singleton object.
@@ -30,27 +30,15 @@ class FilterServer extends Filter
     protected static $instance;
 
     /**
-     * Filters variables passed by Server (Apache SetEnv for instance)
-     * @return FilterServer
+     * Filters variables passed by Post, Get and Cookie.
+     * @return Filter
      */
     public static function getInstance()
     {
         if (!self::$instance) {
-            self::$instance = new self($_SERVER);
-            //$_SERVER = array();		//Too soon to remove the $_SERVER variable. It's being used in lots of places yet.
-            // ¡Lombarte! ¡Lombarte!, ¡Lombarte es cojonudo!, ¡como Lombarte no hay ninguno!
+            self::$instance = new self($_REQUEST);
         }
 
         return self::$instance;
-    }
-
-    /**
-     * Mocks the host for use in scripts.
-     *
-     * @param string $mocked_host
-     */
-    public function setHost($mocked_host)
-    {
-        $this->request['HTTP_HOST'] = $mocked_host;
     }
 }
