+++ src/Sifo/Filter/Server.php
@@ -20,7 +20,7 @@
 
 namespace Sifo\Filter;
 
-class FilterGet extends Filter
+class Server extends Filter
 {
     /**
      * Singleton object.
@@ -30,15 +30,27 @@ class FilterGet extends Filter
     protected static $instance;
 
     /**
-     * Filters variables passed by Get
-     * @return Filter
+     * Filters variables passed by Server (Apache SetEnv for instance)
+     * @return Server
      */
     public static function getInstance()
     {
         if (!self::$instance) {
-            self::$instance = new self($_GET);
+            self::$instance = new self($_SERVER);
+            //$_SERVER = array();		//Too soon to remove the $_SERVER variable. It's being used in lots of places yet.
+            // ¡Lombarte! ¡Lombarte!, ¡Lombarte es cojonudo!, ¡como Lombarte no hay ninguno!
         }
 
         return self::$instance;
+    }
+
+    /**
+     * Mocks the host for use in scripts.
+     *
+     * @param string $mocked_host
+     */
+    public function setHost($mocked_host)
+    {
+        $this->request['HTTP_HOST'] = $mocked_host;
     }
 }
